+++ api/routes/events/[eventId]/instances/index.js
@@ -4,6 +4,7 @@ import { getNextInstance } from "#util/getNextInstance";
 import { z } from "zod";
 import { serializeError } from "#serializeError";
 import { zerialize } from "zodex";
+import { cloneInstanceFromTemplate } from "#util/cloneInstance/cloneInstanceFromTemplate.js";
 
 export const instanceSchema = z.object({
   name: z.string().min(2).max(50),
@@ -11,7 +12,15 @@ export const instanceSchema = z.object({
   endTime: z.string().datetime(),
   startTimeTz: z.string(),
   endTimeTz: z.string(),
-  templateInstanceId: z.string(),
+  templateInstanceId: z.string().optional(),
+
+  locationJobsShifts: z.boolean().default(false),
+  formField: z.boolean().default(false),
+  registrationTier: z.boolean().default(false),
+  registrationPeriod: z.boolean().default(false),
+  registrationPeriodPricing: z.boolean().default(false),
+  upsellItem: z.boolean().default(false),
+  registration: z.boolean().default(false),
 });
 
 export const get = [
@@ -47,11 +56,43 @@ export const post = [
       return res.status(400).json({ message: serializeError(parsed) });
     }
 
+    console.log(parsed.data);
+
     const instance = await prisma.eventInstance.create({
-      data: { ...parsed.data, eventId },
+      data: {
+        eventId,
+        name: parsed.data.name,
+        startTime: parsed.data.startTime,
+        endTime: parsed.data.endTime,
+        startTimeTz: parsed.data.startTimeTz,
+        endTimeTz: parsed.data.endTimeTz,
+      },
     });
 
+    if (parsed.data.templateInstanceId) {
+      const cloneOp = await cloneInstanceFromTemplate({
+        eventId,
+        fromInstanceId: parsed.data.templateInstanceId,
+        newInstance: instance,
+        options: {
+          shiftChildrenByInstanceDelta: true,
+          cloneLocations: parsed.data.locationJobsShifts,
+          cloneJobs: parsed.data.locationJobsShifts,
+          cloneShifts: parsed.data.locationJobsShifts,
+          cloneVolunteerFormFields: parsed.data.formField,
+          cloneRegistrationForms: parsed.data.registration,
+          cloneRegistrationPeriods: parsed.data.registrationPeriod,
+          cloneRegistrationTiers: parsed.data.registrationTier,
+          cloneRegistrationPeriodPricing: parsed.data.registrationPeriodPricing,
+          cloneUpsellItems: parsed.data.upsellItem,
+        },
+      });
+
+      console.log(cloneOp);
+    }
+
     res.json({ instance });
+    // res.json({ message: "Not implemented" });
   },
 ];
