+++ app/components/InstancePicker/InstancePicker.jsx
@@ -1,8 +1,12 @@
 import React from "react";
 import { useInstances } from "../../hooks/useInstances";
-import { DropdownInput, Badge } from "tabler-react-2";
-import { Row } from "../../util/Flex";
+import { DropdownInput, Badge, Typography } from "tabler-react-2";
+import { Col, Row } from "../../util/Flex";
 import { useSelectedInstance } from "../../contexts/SelectedInstanceContext";
+import moment from "moment";
+import { Dropdown } from "../dropdown/Dropdown";
+import { Icon } from "../../util/Icon";
+import { formatDate } from "../tzDateTime/tzDateTime";
 
 export const InstancePicker = ({
   eventId,
@@ -22,7 +26,6 @@ export const InstancePicker = ({
   const { instanceDropdownValue, setInstance } = useSelectedInstance();
 
   const handleChange = (id) => {
-    console.log(id);
     if (id.id === "eventpilot__create") {
       createInstanceInteraction();
     } else {
@@ -31,35 +34,100 @@ export const InstancePicker = ({
     }
   };
 
+  const sorted = [...(instances ?? [])].sort((a, b) => {
+    // Active first
+    if (a.active && !b.active) return -1;
+    if (!a.active && b.active) return 1;
+
+    // Upcoming second
+    if (a.isNext && !b.isNext) return -1;
+    if (!a.isNext && b.isNext) return 1;
+
+    // Past last
+    const aPast = a.endTime && new Date(a.endTime) < new Date();
+    const bPast = b.endTime && new Date(b.endTime) < new Date();
+    if (aPast && !bPast) return 1;
+    if (!aPast && bPast) return -1;
+
+    // Otherwise sort by startTime
+    return new Date(a.startTime ?? 0) - new Date(b.startTime ?? 0);
+  });
+
   const items = [
-    ...(instances ?? []).map((i) => ({
-      id: i.id,
-      label: (
-        <Row gap={1}>
-          {i.name}
-          {i.active && (
-            <Badge color="success" style={{ color: "white" }}>
-              Active
-            </Badge>
-          )}
-          {i.isNext && (
-            <Badge color="primary" style={{ color: "white" }}>
-              Upcoming
-            </Badge>
-          )}
-          {i.endTime && new Date(i.endTime) < new Date() && (
-            <Badge color="orange" style={{ color: "white" }}>
-              Past
-            </Badge>
-          )}
-        </Row>
-      ),
-      searchIndex: `${i.startTime ?? ""} ${i.name ?? ""}`,
-    })),
+    ...(sorted ?? [])
+      .map((i) => ({
+        id: i.id,
+        label: (
+          <Row gap={1}>
+            {i.name}
+            {i.active && (
+              <Badge color="success" style={{ color: "white" }}>
+                Active
+              </Badge>
+            )}
+            {i.isNext && (
+              <Badge color="primary" style={{ color: "white" }}>
+                Upcoming
+              </Badge>
+            )}
+            {i.endTime && new Date(i.endTime) < new Date() && (
+              <Badge color="orange" style={{ color: "white" }}>
+                Past
+              </Badge>
+            )}
+          </Row>
+        ),
+        dropdownText: (
+          <Row gap={1}>
+            <Col align="flex-start">
+              <Row gap={1}>
+                <span style={{ fontWeight: 500 }}>{i.name}</span>
+                {i.active && (
+                  <Badge color="success" style={{ color: "white" }}>
+                    Active
+                  </Badge>
+                )}
+                {i.isNext && (
+                  <Badge color="primary" style={{ color: "white" }}>
+                    Upcoming
+                  </Badge>
+                )}
+                {i.endTime && new Date(i.endTime) < new Date() && (
+                  <Badge color="orange" style={{ color: "white" }}>
+                    Past
+                  </Badge>
+                )}
+              </Row>
+              {i.startTime && (
+                <Row gap={1}>
+                  <Typography.Text
+                    className="text-secondary mb-0"
+                    style={{ fontSize: "0.7rem" }}
+                  >
+                    {formatDate(i.startTime, i.startTimeTz, "M/D/YY")} -{" "}
+                    {formatDate(i.endTime, i.startTimeTz, "M/D/YY")}
+                  </Typography.Text>
+                </Row>
+              )}
+            </Col>
+          </Row>
+        ),
+        searchIndex: `${i.startTime ?? ""} ${i.name ?? ""}`,
+      }))
+      .filter(Boolean),
+
+    showCreate && {
+      type: "divider",
+    },
 
     showCreate && {
       id: "eventpilot__create",
-      label: "Create new instance",
+      label: (
+        <Row gap={1}>
+          <Icon i="plus" size={16} />
+          <span>Create new instance</span>
+        </Row>
+      ),
       searchIndex: "create",
     },
   ].filter(Boolean);
